Restructuring of Code, added json.h header library
Restructured the renderer and vulkan classes to be in a global loader singleton class. Fixed a bug in ECS and in dr librarys. Added json.h from https://github.com/nlohmann/json.

+++ SpriteSparkEngine/Source/Platform/Vulkan/VulkanDescriptors.cpp
@@ -172,7 +172,7 @@ namespace SpriteSpark {
         for (auto& write : m_Writes) {
             write.dstSet = set;
         }
-        vkUpdateDescriptorSets(m_Pool.m_Device.device(), m_Writes.size(), m_Writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(m_Pool.m_Device.device(),static_cast<uint32_t>(m_Writes.size()), m_Writes.data(), 0, nullptr);
     }
 
 }
